refactoring 'school service': renaming/cleanup 'schools.service.js'
