+++ src/services/index.js
@@ -6,7 +6,7 @@ const newsEvents = require('./news/events');
 const content = require('./content');
 const calendar = require('./calendar');
 const ltiTool = require('./ltiTool');
-const school = require('./school');
+const school = require('./school/schools.service');
 const system = require('./system');
 const lesson = require('./lesson');
 const analytics = require('./analytics');
